requirements.txt
attrs==21.4.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
coverage==6.3.2
cryptography==36.0.1
Flask==2.0.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.28
packaging==21.3
pluggy==1.0.0
protobuf==3.19.3
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==7.1.1
pytest-cov==3.0.0
python-dotenv==0.19.2
requests==2.27.1
SQLAlchemy==1.4.31
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.0.2
WTForms==3.0.1
yapf==0.32.0
